web/requirements.txt
dill==0.3.4
django==3.2.5
cassiopeia==4.0.18
python-dotenv==0.18.0
django-cassiopeia==2.1.0
django-extensions==3.1.3
ipython==7.25.0
djongo==1.3.6
role-identification-0.2.0
